[analyzer] fix: move GlobalType to error handler to avoid circular importing
