+++ src/analyzer/error_handler.py
@@ -1,6 +1,18 @@
 from enum import Enum
 from src.lexer.token import Token
 
+class GlobalType(Enum):
+    def __init__(self, name):
+        self._name = name
+    def __str__(self):
+        return self._name
+    def __repr__(self):
+        return self._name
+
+    IDENTIFIER = "IDENTIFIER"
+    FUNCTION = "FUNCTION"
+    CLASS = "CLASS"
+
 class ErrorSrc:
     src = [""]
